fix cache bug

+++ services/crawlers/src/one.ts
@@ -28,7 +28,7 @@ export async function main(slug: string) {
     const shouldSkip = (str: string) => skips.includes(str)
 
     // clear existing scrapes
-    if (!shouldSkip('external') && !shouldSkip('scrapes')) {
+    if (rest.id && !shouldSkip('external') && !shouldSkip('scrapes')) {
       console.log('Clearing existing scrapes for restaurant')
       const res = await db.query(`DELETE FROM scrape WHERE restaurant_id = '${rest.id}';`)
       console.log('Deleted', res.rows)
